+++ packages/flutter/lib/modularity_flutter.dart
@@ -4,4 +4,5 @@ export 'package:modularity_core/modularity_core.dart';
 export 'src/widgets/module_scope.dart';
 export 'src/widgets/module_provider.dart';
 export 'src/widgets/modularity_root.dart';
-export 'src/modularity.dart';
+export 'src/modularity.dart'
+    show Modularity, ModuleLifecycleEvent, ModuleLifecycleLogger;