+++ packages/modularity_test/lib/src/module_test.dart
@@ -33,7 +33,7 @@ Future<void> testModule<T extends Module>(
   );
 
   try {
-    await controller.initialize({});
+    await controller.initialize(<ModuleRegistryKey, ModuleController>{});
     await body(module, testBinder);
   } finally {
     await controller.dispose();